[Refactor]#2 Cell 추상화
